+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/session/room/reporting/DefaultReportingService.kt
@@ -16,17 +16,18 @@
 
 package org.matrix.android.sdk.internal.session.room.reporting
 
-import com.squareup.inject.assisted.Assisted
-import com.squareup.inject.assisted.AssistedInject
+import dagger.assisted.Assisted
+import dagger.assisted.AssistedInject
+import dagger.assisted.AssistedFactory
 import org.matrix.android.sdk.api.session.room.reporting.ReportingService
 
 internal class DefaultReportingService @AssistedInject constructor(@Assisted private val roomId: String,
                                                                    private val reportContentTask: ReportContentTask
 ) : ReportingService {
 
-    @AssistedInject.Factory
+    @AssistedFactory
     interface Factory {
-        fun create(roomId: String): ReportingService
+        fun create(roomId: String): DefaultReportingService
     }
 
     override suspend fun reportContent(eventId: String, score: Int, reason: String) {